+++ src/test/java/com/eventswarm/events/jdo/OrgJsonHttpEventTest.java
@@ -156,21 +156,21 @@ public class OrgJsonHttpEventTest {
     public void integerRetriever() throws Exception {
         JsonEvent.IntegerRetriever retriever = new JsonEvent.IntegerRetriever("a");
         JsonEvent<JSONObject> instance = new OrgJsonHttpEvent(header, new JSONObject("{a:1, b: 2}"), http);
-        assertEquals(1, retriever.getValue(instance));
+        assertEquals(1, (int) retriever.getValue(instance));
     }
 
     @Test
     public void longRetriever() throws Exception {
         JsonEvent.LongRetriever retriever = new JsonEvent.LongRetriever("a");
         JsonEvent<JSONObject> instance = new OrgJsonHttpEvent(header, new JSONObject("{a:1038401717746111341, b: 2}"), http);
-        assertEquals(1038401717746111341L, retriever.getValue(instance));
+        assertEquals(1038401717746111341L, (long) retriever.getValue(instance));
     }
 
     @Test
     public void doubleRetriever() throws Exception {
         JsonEvent.DoubleRetriever retriever = new JsonEvent.DoubleRetriever("b");
         JsonEvent<JSONObject> instance = new OrgJsonHttpEvent(header, new JSONObject("{a:1, b: 3.14159}"), http);
-        assertEquals(3.14159, retriever.getValue(instance));
+        assertEquals(3.14159, (double) retriever.getValue(instance));
     }
 
     @Test
